requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
asyncio==3.4.3
bleak==0.20.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipython==8.29.0
jedi==0.19.2
jiter==0.8.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.7
numpy==1.24.2
openai==1.55.2
packaging==23.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyparsing==3.0.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
stack-data==0.6.3
tqdm==4.65.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
